LayoutTests/imported/w3c/web-platform-tests/webxr/xrRigidTransform_sameObject.https-expected.txt


FAIL XRRigidTransform position and orientation meet [SameObject] requirements assert_equals: XRRigidTransform.matrix returns the same object. expected object "0.04651162773370743,-0.41860464215278625,0.9069767594337463,0,-0.9767441749572754,-0.20930232107639313,-0.04651162773370743,0,0.20930232107639313,-0.8837209343910217,-0.41860464215278625,0,-9,8,-7,1" but got object "0.04651162773370743,-0.41860464215278625,0.9069767594337463,0,-0.9767441749572754,-0.20930232107639313,-0.04651162773370743,0,0.20930232107639313,-0.8837209343910217,-0.41860464215278625,0,-9,8,-7,1"
